% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_spatial_vars.R
\name{compare_spatial_vars}
\alias{compare_spatial_vars}
\title{Creates post-processing output for Atlantis}
\usage{
compare_spatial_vars(
  param.dir,
  run.dirs,
  run.names,
  ref.data,
  init.data,
  out.dir,
  out.name,
  param.ls,
  data.type = "proportion",
  comparison.type = "difference",
  ref.years,
  speciesCodes = NULL,
  plot = T
)
}
\arguments{
\item{param.dir}{string. Path to location of atlantis parameter files}

\item{run.dirs}{character vector of paths to location of atlantis output files}

\item{run.names}{character vector of run names corresponding to run.dirs}

\item{ref.data}{dataframe  that contains reference spatial data with format (polygon|species|var.name|statistic|ref.value)}

\item{init.data}{dataframe that contains initial spatial data  with format (polygon|species|var.name|statistic|init.value)}

\item{out.dir}{string. path to desired location of post-processed output}

\item{out.name}{string. name for output file prefix}

\item{param.ls}{list generated from get_atl_paramfiles()}

\item{data.type}{which type of data is being compared: 'proportion','value'}

\item{comparison.type}{which type of comparison should be made: 'difference','scalar'}

\item{ref.years}{numeric vector with first and last year (years from start) for comparison period}

\item{speciesCodes}{character vector. List of species to make plots for. Defaul = NULL (All species)}

\item{plot}{logical. do you want to generate a plot}
}
\value{
Either saves an R object or returns a list called "result"

Author: Joseph Caracappa
}
\description{
Creates post-processing output for Atlantis
}
