% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_atl_paramfiles.R
\name{get_atl_paramfiles}
\alias{get_atl_paramfiles}
\title{Creates a dataframe of atlantis parameter files based on standard output}
\usage{
get_atl_paramfiles(param.dir, atl.dir, run.prefix, include_catch)
}
\arguments{
\item{param.dir}{string. Path to location of atlantis parameter files}

\item{atl.dir}{string. Path to location of atlantis output files}

\item{run.prefix}{string. Name of the run. filename prefix in the atlantis output}

\item{include_catch}{logical. Whether to include catch output}
}
\value{
list containing parameter file name and full path to file
}
\description{
Creates a dataframe of atlantis parameter files based on standard output
}
