% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_atlantis_diagnostic_figures.R
\name{make_atlantis_diagnostic_figures}
\alias{make_atlantis_diagnostic_figures}
\title{Creates standard diagnostic figures for Atlantis output}
\usage{
make_atlantis_diagnostic_figures(
  out.dir,
  fig.dir,
  atl.dir,
  param.dir,
  run.prefix,
  run.name,
  param.ls,
  benthic.box,
  benthic.level = 4,
  plot.all,
  plot.benthic,
  plot.overall.biomass,
  plot.biomass.timeseries,
  plot.length.age,
  plot.biomass.box,
  plot.c.mum,
  plot.sn.rn,
  plot.recruits,
  plot.numbers.timeseries,
  plot.physics,
  plot.growth.cons,
  plot.cohort,
  plot.diet,
  plot.consumption,
  plot.spatial.biomass,
  plot.spatial.biomass.seasonal,
  plot.catch,
  plot.spatial.catch,
  plot.weight,
  plot.mortality
)
}
\arguments{
\item{out.dir}{string. Path where desired output data should be saved}

\item{fig.dir}{string. Path where desired output figures should be saved}

\item{atl.dir}{string. Path where atlantis model output is located}

\item{param.dir}{string. Path where atlantis parameter files are located}

\item{run.prefix}{sring. Prefix specified in Atlantis.bat file that begins each output file}

\item{run.name}{string. Prefered name to use for output file names}

\item{param.ls}{list. Output from get_atl_paramfiles}

\item{benthic.box}{numeric. Box ID for benthic plots}

\item{benthic.level}{numeric. Level for benthic plots (default is 4 for NEUS model)}

\item{plot.all}{Boolean. Global flag for plotting everything}

\item{plot.benthic}{logical. Benthic plots show timeseries of all benthic and epibenthic groups for one box}

\item{plot.overall.biomass}{logical. Plots showing the total biomass across all functional groups as stacked barplots}

\item{plot.biomass.timeseries}{logical. Plots showing biomass-related timeseries on various aggregations  and reference points}

\item{plot.length.age}{logical. Plots relating to the length-age relationship of age-structured groups}

\item{plot.biomass.box}{logical. Plots relating to biomass but grouped by box}

\item{plot.c.mum}{logical. Plots and tables related to tuning C and mum parameters}

\item{plot.sn.rn}{logical. Plots relating to SN and RN timeseries}

\item{plot.recruits}{logical. Plots of recruitment and SSB timeseries}

\item{plot.numbers.timeseries}{logical. Plots showing timeseries of numbers (as opposed to biomass)}

\item{plot.physics}{logical. Plots of physical statevariables as well as fluxes}

\item{plot.growth.cons}{logical. Plots relating to growth and consumption}

\item{plot.cohort}{logical. Plots showing timeseries of each cohort across age-structured groups}

\item{plot.diet}{logical. Plots showing predation of and consumption by each functional group}

\item{plot.consumption}{Boolean. Plots showing consumption}

\item{plot.spatial.biomass}{logical. Plots showing the spatial (box/level) structure of groups' biomass}

\item{plot.spatial.biomass.seasonal}{logical. Plots showing the spatial (box/level) structure of groups' biomass}

\item{plot.catch}{logical. Plots annual catch(mt) age based catch (numbers) and age based \%ages}

\item{plot.spatial.catch}{logical. Plots showing the spatial (box/level) structure of groups' catch}

\item{plot.weight}{logical. Plots the maximum size of fish in each size class over the domain}

\item{plot.mortality}{logical. Plots Mortality (F, M1, M2) from two output sources (Mort, SpecificMort)}
}
\value{
A series of figures and tables based on output grouping flags

Author: Ryan Morse, modified by J. Caracappa
}
\description{
Used after the post processing routine \code{\link{process_atl_output}}.
Creates the full set of diagnostic and summary figures and tables from atlantis model run.
}
