% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass_box_summary.R
\name{plot_biomass_box_range}
\alias{plot_biomass_box_range}
\title{Plot biomass_box_range}
\usage{
plot_biomass_box_range(
  bio.box,
  bio.box.invert,
  day.min = NA,
  day.max = NA,
  plot.presence = T,
  species.list = NULL,
  save.fig = T,
  fig.dir
)
}
\description{
Plot biomass_box_range
}
