% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass_box_summary.R
\name{plot_biomass_box_season}
\alias{plot_biomass_box_season}
\title{Functions to make box/epu species biomass presence/absence}
\usage{
plot_biomass_box_season(
  bio.box,
  bio.box.invert,
  plot.presence = F,
  species.list = NULL,
  save.fig = T,
  fig.dir,
  tolerance = 0.1
)
}
\description{
Functions to make box/epu species biomass presence/absence
}
