% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_length_freq.R
\name{plot_length_freq}
\alias{plot_length_freq}
\title{Creates plots comparing the length frequency distribution of model catch and biomass against data}
\usage{
plot_length_freq(
  param.dir,
  run.dir,
  run.name,
  catch.ref,
  survey.ref,
  ref.years,
  bin.width = 1,
  out.dir,
  out.name,
  param.ls,
  speciesCodes = NULL,
  fleetCodes = NULL
)
}
\arguments{
\item{param.dir}{string. Path to location of atlantis parameter files}

\item{run.dir}{character vector of path to location of atlantis output files}

\item{run.name}{character vector of run name corresponding to run.dir}

\item{catch.ref}{dataframe  that contains reference catch data with format (Code|len.lower|len.upper|cum.prob)}

\item{survey.ref}{dataframe that contains reference biomass data  with format (Code|len.lower|len.upper|cum.prob)}

\item{ref.years}{numeric vector with first and last year (years from start) for comparison period}

\item{bin.width}{numeric. Width of bins (in cms) in which to bin Fish length. Default = 1}

\item{out.dir}{string. path to desired location of post-processed output}

\item{out.name}{string. name for output file prefix}

\item{param.ls}{list generated from get_atl_paramfiles()}

\item{speciesCodes}{character vector. List of species to make plots for. Default = NULL (All species)}

\item{fleetCodes}{character vector. List of fleets to make plots for., Default = NULL (All fleets)}
}
\value{
PDF of all fleet x species combination cumulative length frequency distributions

Author: Joseph Caracappa
}
\description{
Creates plots comparing the length frequency distribution of model catch and biomass against data
}
