% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_atl_output.R
\name{process_atl_output}
\alias{process_atl_output}
\title{Creates post-processing output for Atlantis}
\usage{
process_atl_output(
  param.dir,
  atl.dir,
  out.dir = file.path(atl.dir, "Post_Processed/Data/"),
  run.prefix,
  param.ls,
  agg.scale = "day",
  save.out,
  large.file = F,
  system,
  process.all = F,
  plot.all = F,
  plot.benthic = F,
  plot.overall.biomass = F,
  plot.biomass.timeseries = F,
  plot.length.age = F,
  plot.biomass.box = F,
  plot.c.mum = F,
  plot.sn.rn = F,
  plot.recruits = F,
  plot.numbers.timeseries = F,
  plot.physics = F,
  plot.growth.cons = F,
  plot.cohort = F,
  plot.diet = F,
  plot.consumption = F,
  plot.spatial.biomass = F,
  plot.spatial.biomass.seasonal = F,
  plot.LTL = F,
  plot.catch = F,
  plot.mortality = F,
  plot.max.weight = F,
  plot.spatial.overlap = F
)
}
\arguments{
\item{param.dir}{string. Path to location of atlantis parameter files}

\item{atl.dir}{string. path to location of atlantis output files}

\item{out.dir}{string. path to desired location of post-processed output}

\item{run.prefix}{string. Prefix for atlantis run output (specified in runcommand.bat)}

\item{agg.scale}{Scale to aggregate dietcheck biomass from (either 'raw','month', or 'year' )}

\item{param.s}{list generated from get_atl_paramfiles()}
}
\value{
Either saves an R object or returns a list called "result"

Author: Ryan Morse, modified by Joseph Caracappa
}
\description{
Creates post-processing output for Atlantis
}
