% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_catch_fleet.R
\name{process_catch_fleet}
\alias{process_catch_fleet}
\title{Creates post-processing output for Atlantis}
\usage{
process_catch_fleet(fishery.prm, catch, groups.file)
}
\arguments{
\item{fishery.prm}{string. Path to location of atlantis fishery csv file}

\item{catch}{string. path to location of atlantis catch output file}

\item{groups.file}{string. path to location of atlantis functional groups csv file}
}
\value{
returns a dataframe

Author: Joseph Caracappa
}
\description{
Creates post-processing output for Atlantis
}
